Refactor test imports and remove unused code, del migration files
