+++ backend/entities/views.py
@@ -2,6 +2,7 @@ from django.shortcuts import get_object_or_404
 from rest_framework import status, viewsets
 from rest_framework.request import Request
 from rest_framework.response import Response
+from rest_framework.throttling import AnonRateThrottle, UserRateThrottle
 
 from backend.paginator import CustomPagination
 
@@ -35,7 +36,6 @@ from .serializers import (
     OrganizationTaskSerializer,
     OrganizationTopicSerializer,
 )
-from rest_framework.throttling import AnonRateThrottle, UserRateThrottle
 
 
 class OrganizationViewSet(viewsets.ModelViewSet[Organization]):
