+++ backend/events/tests.py
@@ -1,10 +1,8 @@
 from django.test import TestCase
 from django.urls import reverse
-from tests.test_base import BaseTestThrottle
+from tests.throttle import BaseTestThrottle
 
 
 class EventsThrottleTest(BaseTestThrottle):
     __test__ = True
     url = reverse("events:event-list")
-    anon_throttle = "7/min"
-    user_throttle = "10/min"
