--- backend/tests/__init__.py
@@ -1,1 +1,0 @@
-from tests.test_base import BaseTestThrottle
